Drop the Ruby app
We are about to redo the site as Hugo site, and do not want to serve
anything dynamically via Ruby anymore.

Note: This commit leaves some files behind that are part of the Rails
App. The reason is that these parts will be moved, renamed and adjusted
later on, as they are still needed (for example to render the book, even
if the result is not put into a database ase with the Rails App but
instead will be committed to the Git repository).

Signed-off-by: Matt Burke <spraints@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>