Convert 'About' pages to Hugo
This changes the pages so that they render correctly using Hugo.

Note that this commit makes it look as if the
`app/views/about/index.html.erb` file was simply deleted. However, its
content was adapted and merged into `layouts/_default/baseof.html`.
Since Git's diff machinery operates only on single files or at most a
pair of file names in case of a rename, there is not really any good way
to present this movement of one file's content into an already-existing
other file.

Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>