Use relative URLs
To be able to develop git-scm.com changes in a personal fork (where the
URL is of the form `https://user.github.io/git-scm.com/`, i.e. the site
is _not_ located at the root directory) as well as reuse the result in
the main repository (where the site does live in the root directory), it
is highly recommended to use relative URLs where possible.

This is particularly necessary because we want
https://git.github.io/git-scm.com/ to render correctly, too.

This trick is a bit tricky because Hugo's `relURL` function is only
available in the `layouts/` directory, and our custom `relurl` shortcode
has to be used inside `content/`, and the trick was performed by:

  sed -i 's,\(href=\|src=\)["'\'']/\([^{"'\'']*\)["'\''],\1"{{< relurl "\2" >}}",g' \
    $(git ls-files content/\*.html)

  sed -i 's,\(href=\|src=\|action=\)["'\'']/\([^{"'\'']*\)["'\''],\1"{{ relURL "\2" }}",g' \
    $(git ls-files ':^content/' \*.html)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>