update-manual-pages: optionally force a complete rebuild
Sometimes a bug is found in the way the manual pages are built,
requiring a change e.g. in the `update-docs.rb` script or in the
corresponding style sheets.

In such a case, the manual pages have to be rebuilt from scratch, even
if they exist already.

This commit introduces the `force-rebuild` flag to do this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>