ci: verify the order of the search results
We definitely want the search results for, say, "commit" to show the
manual page of that command first.

This took quite a bit of work to accomplish, including a change to
Pagefind itself (https://github.com/CloudCannon/pagefind/pull/534) as
well as some extensive fiddling with the ranking weights (including, I
kid you not, running a Powell's optimization of the weights).

To ensure that the results are shown in the desired order, let's add an
explicit check for that in the PR builds.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>