amend! book: also generate the URLs for the downloadable formats of the book
book: also generate the URLs for the downloadable formats of the book

... but do this only for the English version, as the other versions are
not available in downloadable formats.

As the idea is to run this script in a GitHub workflow on a shallow
clone, let's be prepared for the situation where the tip commit of the
default branch is _not_ (yet?) tagged and therefore no tag is present in
the clone. In this situation, let's just go ahead and fetch all tags,
then move along and update the URLs accordingly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>